#10334: Migrate abs_bw to TTNN with overloading
