+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -635,8 +635,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.prod_bw
 
-.. autofunction:: tt_lib.tensor.complex_abs_bw
-
 .. autofunction:: tt_lib.tensor.complex_recip_bw
 
 .. autofunction:: tt_lib.tensor.complex_mul_bw
