+++ tt_eager/tt_dnn/op_library/complex/complex_ops.cpp
@@ -297,21 +297,6 @@ std::vector<ComplexTensor> complex_div_bw(const ComplexTensor& grad, const Compl
     return grad_tensor;
 }
 
-// complex abs
-// self: grad * self.sgn()
-std::vector<ComplexTensor> complex_abs_bw(const Tensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config) {
-    std::vector<ComplexTensor> grad_tensor;
-    Tensor result = ttnn::operations::complex_unary::_abs(input, output_mem_config);
-    Tensor grad_inp_r = where(ttnn::eqz(result, output_mem_config), zeros_like(result, output_mem_config), ttnn::multiply(grad, ttnn::multiply(input.real(), ttnn::reciprocal(result, output_mem_config), std::nullopt, output_mem_config),std::nullopt, output_mem_config), output_mem_config );
-    Tensor grad_inp_i = where(ttnn::eqz(result, output_mem_config), zeros_like(result, output_mem_config), ttnn::multiply(grad, ttnn::multiply(input.imag(), ttnn::reciprocal(result, output_mem_config), std::nullopt, output_mem_config),std::nullopt, output_mem_config), output_mem_config );
-    ComplexTensor grad_inp = ComplexTensor({ grad_inp_r, grad_inp_i});
-    result.deallocate();
-    grad_inp_r.deallocate();
-    grad_inp_i.deallocate();
-    grad_tensor.emplace_back(grad_inp);
-    return grad_tensor;
-}
-
 // complex reciprocal
 // self: -grad * (result * result).conj()
 std::vector<ComplexTensor> complex_recip_bw(const ComplexTensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config) {
